strip env_config deps
