+++ config/config_test.go
@@ -22,10 +22,12 @@ import (
 	"os"
 	"path"
 	"testing"
+	"time"
 
 	"github.com/apigee/apigee-remote-service-envoy/v2/testutil"
 	"github.com/apigee/apigee-remote-service-envoy/v2/util"
 	"github.com/apigee/apigee-remote-service-golib/v2/errorset"
+	"github.com/google/go-cmp/cmp"
 	"gopkg.in/yaml.v3"
 )
 
@@ -633,6 +635,103 @@ func TestLoadFromEnvironmentVariables(t *testing.T) {
 	c = Default()
 	if err := c.Load(tf.Name(), "", DefaultAnalyticsSecretPath, true); err == nil {
 		t.Errorf("c.Load() should have given error on bad jwks")
+	}
+}
+
+func TestLoadEnvironmentSpecs(t *testing.T) {
+	tests := []struct {
+		desc        string
+		filename    string
+		wantEnvSpec EnvironmentSpec
+	}{
+		{
+			desc:     "good config file with references to env config files",
+			filename: "./testdata/good_config.yaml",
+			wantEnvSpec: EnvironmentSpec{
+				ID: "good-env-config",
+				APIs: []APISpec{
+					{
+						BasePath: "/v1",
+						Authentication: AuthenticationRequirement{
+							Requirements: JWTAuthentication{
+								Name:       "foo",
+								Issuer:     "bar",
+								JWKSSource: RemoteJWKS{URL: "url", CacheDuration: time.Hour},
+								In:         []APIOperationParameter{{Match: Header("header")}},
+							},
+						},
+						ConsumerAuthorization: ConsumerAuthorization{
+							In: []APIOperationParameter{{Match: Header("x-api-key")}},
+						},
+						Operations: []APIOperation{
+							{
+								Name: "op-1",
+								HTTPMatches: []HTTPMatch{
+									{
+										PathTemplate: "/petstore",
+										Method:       "GET",
+									},
+								},
+							},
+							{
+								Name: "op-2",
+								HTTPMatches: []HTTPMatch{
+									{
+										PathTemplate: "/bookshop",
+										Method:       "POST",
+									},
+								},
+							},
+						},
+						HTTPRequestTransforms: HTTPRequestTransformations{
+							SetHeaders: map[string]string{
+								"x-apigee-route": "route",
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.desc, func(t *testing.T) {
+			c := &Config{}
+			if err := c.Load(test.filename, "", "", false); err != nil {
+				t.Errorf("c.Load() returns unexpected: %v", err)
+			}
+			if l := len(c.EnvironmentSpecs.Inline); l != 1 {
+				t.Fatalf("c.Load() results in %d EnvironmentSpec, wanted 1", l)
+			}
+			if diff := cmp.Diff(test.wantEnvSpec, c.EnvironmentSpecs.Inline[0]); diff != "" {
+				t.Errorf("c.Load() results in unexpected EnvironmentSpec diff (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestLoadEnvironmentSpecsError(t *testing.T) {
+	tests := []struct {
+		desc     string
+		filename string
+	}{
+		{
+			desc:     "bad env config files",
+			filename: "./testdata/bad_config_1.yaml",
+		},
+		{
+			desc:     "non-existent env config files",
+			filename: "./testdata/bad_config_2.yaml",
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.desc, func(t *testing.T) {
+			c := &Config{}
+			if err := c.Load(test.filename, "", "", false); err == nil {
+				t.Errorf("c.Load() returns no error, should have got error")
+			}
+		})
 	}
 }
